pull up backend independent part of Legion context policy to context
